new cache type fix download progress check free space before downloading logging game download errors

+++ InAppStorySdk/inappstorysdk/src/main/java/com/inappstory/sdk/stories/cache/FileLoadProgressCallback.java
@@ -3,7 +3,7 @@ package com.inappstory.sdk.stories.cache;
 import java.io.File;
 
 public interface FileLoadProgressCallback {
-    void onProgress(int loadedSize, int totalSize);
+    void onProgress(long loadedSize, long totalSize);
     void onSuccess(File file);
     void onError();
 }
