+++ InAppStorySdk/inappstorysdk/src/main/java/com/inappstory/sdk/stories/ui/ugclist/UgcStoryListItem.java
@@ -133,7 +133,7 @@ public class UgcStoryListItem extends BaseStoryListItem {
         if (InAppStoryService.isNull()) return;
         Downloader.downloadFileBackground(url, InAppStoryService.getInstance().getFastCache(), new FileLoadProgressCallback() {
             @Override
-            public void onProgress(int loadedSize, int totalSize) {
+            public void onProgress(long loadedSize, long totalSize) {
 
             }
