+++ InAppStorySdk/inappstorysdk/src/main/java/com/inappstory/sdk/utils/ZipLoadCallback.java
@@ -3,5 +3,5 @@ package com.inappstory.sdk.utils;
 public interface ZipLoadCallback {
     void onLoad(String baseUrl, String data);
     void onError();
-    void onProgress(int loadedSize, int totalSize);
+    void onProgress(long loadedSize, long totalSize);
 }
